requirement.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
requests>=2.31.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.112.0
